rename ellipse to circle and remove unused style
