+++ ui/DesignSystem/Primitive/Icon/Circle.svelte
