--- ui/DesignSystem/Primitive/Icon/Ellipse.svelte
@@ -1,18 +1,0 @@
-<script>
-  export let style = null;
-</script>
-
-<svg
-  {style}
-  width="16"
-  height="16"
-  viewBox="0 0 16 16"
-  fill="var(--color-foreground-level-5)"
-  xmlns="http://www.w3.org/2000/svg">
-  <path
-    fill-rule="evenodd"
-    clip-rule="evenodd"
-    d="M8 10C9.10457 10 10 9.10457 10 8C10 6.89543 9.10457 6 8 6C6.89543 6 6
-    6.89543 6 8C6 9.10457 6.89543 10 8 10ZM8 13C10.7614 13 13 10.7614 13 8C13
-    5.23858 10.7614 3 8 3C5.23858 3 3 5.23858 3 8C3 10.7614 5.23858 13 8 13Z" />
-</svg>
